DPC-74: Endpoint now returns a 204 status, along with a Content-Location header of the fake job.
This still needs to be improved as it's using hardcoded values and allows for any parameters to be passed.
We've also improved some of the testing for both the capabilities statement as well as the resources.

The maven test integration is wonky, it's only running the Capabilities test, even though IntelliJ runs everything.