DPC-91: Attribution Service was incorrectly using the Local Attribution Service
The attribution engine was not using the correct database backend. This fixes the injection so that the tests are no longer passing (which is the correct behavior for right now).

Getting Jersey/Guice/Hibernate to play nicely together is a bit of a challenge. We have to wrap the DB Resources in special proxy handling, in order to get the necessary error handling to work.