DPC-114: Fixed failing database loading
When running as a unified JAR, the database configuration was failing to load due to the wrong version of javassit being loaded. The solution was to move the dependency declaration to the root pom.

This also addresses DPC-96, and by ‘addresses’ I mean, we’re ignoring it. There’s some complex interaction going on that’s causing the database configuration to ‘fail’, but actually load. The solution, for now, is to simply catch the exception, make sure it’s ‘expected’ and if not, rethrow.