DPC-114: Fix unit and integration tests to all pass when run as a single unit (#16)
* DPC-114: Fixed failing database loading

When running as a unified JAR, the database configuration was failing to load due to the wrong version of javassit being loaded. The solution was to move the dependency declaration to the root pom.

This also addresses DPC-96, and by ‘addresses’ I mean, we’re ignoring it. There’s some complex interaction going on that’s causing the database configuration to ‘fail’, but actually load. The solution, for now, is to simply catch the exception, make sure it’s ‘expected’ and if not, rethrow.

* DPC-114: Fixes failing integration tests

These tests were failing because they haven’t been run in a while. Some small fixes and they’re good to go.

* DPC-114: Modified BBClient constructor improve keystore handliing

The BBClient can now read the keystore from the resources, provided the relevant configuration key is not set.

Eventually, this should be injected by Guice.

* DPC-114: Code review fixes to remove superfluous logging

Removed the logging statement which was printing out the DB connection details.

Also cleaned up some typos in the Javadocs.

* DPC-114: Improved error logging in constructor

When the class is constructed by Guice, it swallows the core error and wraps it in a general exception error, which is difficult to read.

This change directly logs the error before Guice throws its nonsense.