DPC-127: Moved vendored resources to Jitpack
We can use Jitpack for any dependencies that are open source, which means we can pull in our forks for type safe config and HAPI. Anytime we update one of our vendored resources, it will automatically rebuild and install, no user intervention required.

So we don’t need the submodules anymore.