DPC-20: Cleaned up attribution tests  to avoid random failures
If the attribution tests failed, the cleanup steps were not being run, which would lead to failures on the next test execution as well.

By moving the truncate command into the setup method, we can be confident that each test will start with a clean slate, even if the previous test didn’t clean up correctly.