DPC-162: Modified test order to try and ensure we can clean up database connections
A few small tweaks to the tests to try and get things working smoother.
DPC-172 has been created to address an issue where Postgres connects are not getting cleaned up correctly.
I don’t think this is a blocking issue because we only seed the database once and then using the existing pool for all future connections.