DPC-169: Implement typesafe SQL (#35)
Added support for the JOOQ SQL framework, which means we now have a clean, typesafe interface for executing custom SQL queries.

This is primarily used in the SeedCommand class, since we can't inject the Hibernate context in a Command because it's not configured yet. This lets us remove a bunch of code and much more cleanly express our intent.

The Java classes are generated by JOOQ, based on our JPA entities. Due to the order of compilation, we had to move the entities in the dpc-common module, so that they would be available for JOOQ to parse. This add a step to the generate-sources phase of maven, but it should be handled automatically. The resulting classes are generated in the target/ directory, so there's nothing for us to commit to git.