DPC-22: Add task to automatically prune attribution relationships (#36)
Added a cron job to automatically expire attribution relationships that were created before a given expiration threshold. The job is run and managed by Dropwizard using the Sundial job scheduler. It's currently configured to run everyday at midnight (based on the system clock).

The testing for it is rather simplistic, we essentially seed the database with attributions and manually override the creation time to be before the expiration threshold. DPC-186 has been created to improve the testing process by manually creating attributions using the current date time and verifying that they are not removed.

In the future, we may want to consider moving the JOOQ SQL into a stored procedure and potentially schedule it within the database itself, but we'll cross that bridge when we get to it.

This PR also includes some changes to the SeedCommand, by using JOOQ we've made things significantly simpler and safer.