DPC-80: Implement Redis based Job Queue  (#39)
This implements a Redis based job queue, in order to fully decouple the API service from the Aggregation Service, as proposed in DPC-80. Because this queue lives inside the host services, DPC-154 is completed as well, since it deploys along side the api and aggregation services.

This simplifies both the deployment and operational footprint by relying on services that we already have deployed (or will deploy) in the environments.
Redis serves as the main messaging layer between the two services, with Postgres serving as the persistence layer.

When a job is added to the queue, it's persisted in Postgres and the jobID is written to the Redis queue.
On the consuming side, the aggregation service monitors for jobs in the queue and then fetches the record from Postgres in order to get the job arguments. If the job is still in the correct state (e.g. QUEUED) it's moved into the RUNNING state and the job is processed.
Once the job is done, it is completed by setting the status to either COMPLETED or FAILED.
The API service can poll the queue for the job status, which is fetched from the Postgres table.

Unit tests have been implemented to check basic functionality and the functional E2E tests are passing as well.

While this queue fulfills the primary design goals, several improvements still need to be made:

It's possible for the queue to enter an inconsistent state where the queue and/or the database fail to persist correctly, leaving the job in only one service, but not the other. We either need to implement some locking or improve our rollback logic. We may also want to consider moving the update logic into the database, doing an update, only if the job is in the correct state.

The configuration for the redis queue is messy and spread out across the various service files. DPC-94 should help with this, but for now, it's pretty brittle.

The database migrate is currently handled by Hibernate itself, which is incredibly simplistic and unreliable, we'll need to implement some more robust schema migration using Liquibase.

Given that the proposed architecture uses Redis for caching patient data, we'll need to separate out the instances to ensure that no patient data crosses the data/app boundary.

We need to implement cleanup logic to purge completed jobs from the database tables, as well as handle any stale jobs that were not completed correctly.

When a job fails, it's not re-added to the queue, that might be a good idea for transient failures (e.g. IO or BB failures).

The job failure currently passes no failure messages, just the pass/fail status.