DPC-23: Add ability to resubmit roster (#43)
* DPC-23: Added Jooq based roster engine

Moving from Hibernate to Jooq significantly reduces the complexity and code size, and should provide a more flexible update policy going forward.

Hibernate still has value though, so I’m not sure I’ll pull it out of the engine yet.

* DPC-23: Updated unique constraints and upsert process

Upserting a record is now handled by a custom upserter class, that allows us to control which fields DO NOT get updated and under what conditions.

* DPC-23: Updated SeedCommand to utilize a shared DB helper method

The SeedCommand and Roster Engine now use a shared DB helper method, which reduces the amount of duplicated code.

Unit tests are passing up to removing attribution relationships.

* DPC-23: Unit tests are now passing.

Attribution tests are now passing to the same level that they were with the Hibernate backend.

* DPC-23: We can now add patients to existing rosters

* DPC-23: Skip javadoc generation for generated sources

We shouldn’t create Javadocs for JOOQ sources, this also fixes a bad javadoc in the Abstract upserter.

* DPC-23: Removed Hibernate annotations

We don’t need them anymore.

Also added some basic error checking to determine if updates/deletions failed.

* DPC-23: Updated Travis to use Postgres 9.6

We need at least Postgres 9.5 in order to use the ON CONFLICT statement, so Travis is upgraded and a note has been added in the README.

* DPC-23: Fixed attribution removal

We only want to remove the Attribution relationship itself, not the patient or provider records.

* DPC-23: Address null values in logging statements

The upsert logic only returns the record ID, so we can’t use it to get the provider/patient ID as well. No problem, we can fix that.

* DPC-23: Migrated DB timestamps to Java types

Removed our usage of the old java.sql.* timestamps and moved to the Java 8 types. This required a changed to Jooq  to have it automatically generated the OffsetDateTime type.

Also renamed the handleAttributionBundle method to be more descriptive of what it actually does.

* DPC-23: Refactored attribution tests

Split the attribution FHIR test into two pieces, which makes it easy to add a third when we land DPC-209.