DPC-216: Fix failing AWS tests (#48)
* DPC-216: Add toString override to JobModel

* Modify Travis script to catch test failures

Unify all build/test steps into a single script.

* DPC-216: Added no arg constructors to JobCompletionModel

Added no-arg constructors to the JobCompletionModel, so that Jackson can de serialize correctly.

Fixed an indentation error with the Jacoco test coverage script.

Updated EndToEndTest to account for multiple resources being created by the export job. We can now test each one and ensure that it’s correct.

* DPC-216: Fix EOB resource checking

Generified the resource checking function. We should not the EOB resources are actually Bundles, so we need to check for the resource type, but cast to something different.

Travis should actually run tests now.

* DPC-221: Updated docker tags and repositories for AWS

When building, we now create the images using the ECR repository and build each image as a unique tag.

* DPC-216: Improved logging of Job status and outcomes

Added some additional logging so that we can monitor job progress more effectively.

* DPC-216: Add sandbox config files.

* DPC-216: Update demo command to provide host option

We can now specify which API host to execute the demo command against.