DPC-215: Fix Hibernate cascade error
I broke the Hibernate cascade setting, but it’s now fixed. This was causing the JobResult model to not get updated on completion.

Also moved the JSON output logging to the trace level.

Small tweak to the EndToEndRequest test to have it evaluate all of the properties on the JobCompletionModel in one fell swoop.