DPC-215: Improve Application Healthchecks (#49)
* DPC-215: Added Attribution and Aggregation service healthchecks

Added some health checks to the Aggregation service.
Attribution service already has the necessary checks enabled, all that’s required it to expose the admin port in the docker-compose file.

* DPC-215: Removed mock queues

We’ve centralized on using the in-memory queue for testing, so these classes are no longer needed.

* DPC-215: Redis queue check.

Added the ability to check the health status of the Job Queue

* DPC-215: Added API health checks

The API now checks that the Job Queue and the Attribution Service are healthy.

* DPC-215: Moved healthy endpoint to correct service.

* DPC-215: Address code review comments

* DPC-215: Fix failing integration tests

The Truncate command URI was incorrect.

* DPC-215: Sessions are no longer injected

We no longer directly inject Session resources, instead, we hand off the SessionFactory and rely on the services to open/close their own sessions.

* DPC-215: Attribution Engine is now injected

We’re binding against the interface rather than the concrete implementation, which is the right way to do things.

* DPC-215: No longer mapping aggregation engine port

We don’t need to expose it and it makes things more complicated.

* DPC-215: Address code review comments

Renamed isHealthy method to assertHealthy to make it more clear what the return semantics are.

* DPC-215: Add RosterEngineHealthcheck

The Roster engine healthcheck was never wired into the system, that has been resolved.

I also modified the AttributionEngine health checks to only return the throwable message on error, not the entire throwable, which is completely unhelpful. Instead, we log it.

* DPC-215: No longer directly injecting DSL context

We’re not longer directly injecting the Jooq DSLContext, because we weren’t always cleaning it up.

Now we have to manually create a context each time we use it, but at least we’re not leaking resources.

* DPC-215: Address error prone warnings

Error prone is building clean now.

* DPC-215: Fix attribution health check URL

Added a missing underscore to the AttributionServiceClient

* DPC-215: Refactored Queue test and added BB health check test

Trying to improve the diff coverage, refactored the JobQueue check into the queue module, so it’s injected by default.

Added a new test for the BB client health check.

* DPC-215: Improve test coverage and method naming

Renamed isHealthy to assertHealthy in the JobQueue.

Added some unit tests to verify that the exceptions are bubbled up to the health checks themselves.

* DPC-215: Address ErrorProne warnings and failing tests

Cleaned up a couple of issues reported by ErrorProne.

Updated iterable subscriber to dispose of itself when the subscription completes, otherwise we’ll leak resources.

Modified the JobModel to fetch the results eagerly, so we can close the Hibernate session appropriately. Also, we’re now only cascading persist operations, that way we avoid updating duplicate JobResults on each merge operation.

* DPC-215: Fix failing integration test

The integration test was failing due to the job updates not being persisted correctly.

This adds a call to clear() to empty the Hibernate session cache before doing the updates.

Also added an error handler to the main job subscription.

* DPC-215: Fix Hibernate cascade error

I broke the Hibernate cascade setting, but it’s now fixed. This was causing the JobResult model to not get updated on completion.

Also moved the JSON output logging to the trace level.

Small tweak to the EndToEndRequest test to have it evaluate all of the properties on the JobCompletionModel in one fell swoop.