DPC-13: Organization Registration (#60)
* DPC-13: Added command for registering organizations

Added simple command to input an organization resource and send it to the Attribution Service.

The corresponding API endpoint does not exist, but the integration and parsing works.

* DPC-13: Initial commit of Organization endpoint for registration

Added OrganizationEntity for storing FHIR resource in database.

Created a new FHIRConvertable interface for adding to/from FHIR methods to a given resource.

* DPC-13:  Entity serialization is working correctly.

We can now create an OrganizationEntity from the underlying FHIR resource.

The extractors are pretty clunky, but the basic logic is working.

* DPC-13: Organization registration command is working

A couple of small tweaks to get organization registration working.

* DPC-13: Expanded org resource test

Added a negative test to the organization resource endpoint.

Modified travis file to clean install as the first step, to avoid any caching issues.

Registration Command now outputs to the stderr when there's a problem.

* DPC-13: Normalized Organization FHIR resource

Organization resource is now persisted across a number of tables, still a bunch of empty values, but we're getting closer.

Added a simple converter system for translating between FHIR and Hibernate entity types.

* DPC-13: Entity references are working

Yup, it works now.

squash

* DPC-13: Organization Registration is now working

We can register an organization, with its contacts and endpoints in a single call.

This should only be available when connected to CMSNet, as we don't publicly map the ports for the attribution service. Still, in the future, we might consider moving this to the admin layer.

* DPC-13: Fix failing compilation and integration tests

Cleaned up some of the attribution tests and addressed some ErrorProne warnings.

* DPC-13: More error prone fixes

Cleaned up a couple of ErrorProne warnings, also modified the config to skip Jooq generated sources, which we don't have any control over.

* DPC-13: Address CodeClimate errors.

A couple of small tweaks to make CodeClimate happy.

* DPC-13: Refactored organization test resources

We now have one organization tmpl, which we can use for both testing and creating new orgs.

* DPC-13: Fix pom typo

Bad typo in pom broke the build, fixed.

* DPC-13: Fix resource typo

Pointed to the wrong directories. Embarrassing, but fixed.

* DPC-13: Address code review comments.

Addressed comments from code review