DPC-125: Refactor Capabilities Statement
Refactor the capabilities statement to accurately reflect the Rest endpoints.

We now correctly nest the REST resources. Unfortunately, there's no way to tag operations on resources as part of STU3, so we have to leave it off for now. We also need to find a way to indicate our Job and Data endpoints, since they're not actually part of our FHIR implementation.