DPC-30: Add token authentication and validation (#81)
* DPC-305: Initial wiring of Organization resource

Added organization resource to API frontend/

Nothing is wired up yet, but the skeleton is there.
We need to integrate the auth handling, in order to get the organization injected correctly.

* DPC-30: Initial commit of authz/authn filters

* DPC-30: Attribution service has Organization search endpoint

We now have the ability to search for an Organization with a given token.

We'll need to refactor the OrganizationEntity model to allow the token IDs to be queried directly.

* DPC-30: Token lookup is working

We can now lookup an Organization based on its token, which we can then validate and ensure is working correctly.

Still need to add FHIR generation

* DPC-30: Organization components are FHIRable

We can now convert our organization related entities into FHIR resources.

It's really clunky and not very composable, but it works for now. We'll need to revisit if we expand our supported data model.

* DPC-309: Added synthetic resource bundles

Used the syntheticMass project to generate some fake data that we can use for seeding the attribution database.

Added a small README file which contains the details of what changed.

* DPC-309: Organization and Endpoint resources are seeding correctly.

A couple of changes to the serializers/deserializers means we can now persist Organizations and their Endpoints correctly.

* DPC-308: Providers are now seeded correctly.

Well, that was easy.

* DPC-308: Reconcile changes with existing tests

We need to remove the autogenerated ID from the EndpointEntity, otherwise any manual values get overwritten by Hibernate and errors are thrown.

Also address an NPI when an Endpoint doesn't have a managing organization, which is fine in some cases.

* DPC-30: Authorization and Authentication handlers are currently wired up

We're now able to annotate protected endpoints and verify that the correct organization is making the request.

* DPC-30: Added simple javadocs

Added some documentation to the newly created classes.

* DPC-30: Refactored authorization to be path based

We can now directly check that a Macaroon is valid for a given Organization.

This means we don't have to pull the entire org during the auth phase, we can just do a simple GET request to the Attribution Service.

* DPC-30: Attribution resource can now return specific organizations

Added READ support to Organization endpoint.

* DPC-30: Fixed incorrect handling of Macaroon caveats

Our custom verifiers weren't correctly handling non-matching caveats.

I've updated the flow logic to ensure that success is the exceptional case, to avoid these issues in the future.
Also created a new set of unit tests that verify any validators that we create ourselves, work correctly.

* DPC-30: Token validation now returns a proper HTTP response

The attribution service returns either a 200 or a 401 based on whether or not the provided token is valid.

* DPC-30: Disabled Dropwizard Auth handlers

We no longer need to use the built in Dropwizard authorization and authentication handlers.

We can simply call the correct verification endpoint and the Attribution Service will take care of everything for us.

* DPC-309: Refactor SeedCommand run method

Moved the individual database operations into their own helper methods.

* DPC-309: Code climate cleanups

* DPC-30: Code climate cleanups

* DPC-30: Refactored Auth test into helper class

Moved some of test utils into a helper class. This will be more useful for the work in DPC-308.

* Fix bad merge

Merging master went awry. This should fix some things.

* Uppercased Macaroons tables name

For some reason, lowercase table names don't seem to work with Liquibase, but making them uppercase ensures they get processed correctly.

* Add missing provider annotation

* DPC-30: Renamed authentication config option

Rename the new configuration variable to match what it's actually supposed to be.

* DPC-30: Refactored auth test into unit and integration

We now have two authorization tests, one to verify the annotations themselves and the other verify that everything works as expected.

* DPC-30: Re-add authentication integration test

This should not have been removed. Oops.