DPC-371: PathAuthorizer is more generic
Refactored the authentication/authorization logic to allow more generic resource associating for a given Organization.

Right now, we can look for generic Resource IDs and verify that the Organization has some association to the given resource. We rely on the methods in the Attribution service to do the actual associations.