DPC-303: Refactored authentication handlers into a factory pattern
Given that we can't pass state between the authenticators and the authorizers, we need to refactor the handlers to create new classes on each auth request.

This is fairly easy to do with Guice but takes a lot of manual switching up of signatures and tests.
The PractitionerResource and AuthenticationTests are passing, but the rest of the suite is mostly disabled.