DPC-303: Link providers to organizations (#93)
* DPC-303: Initial commit of ProviderRoleResource

Added the basic plumbing for creating provider roles. Also updated the DB migration scripts.

* DPC-303: Refactored test helpers

Moved a couple of utilities into a helper class.

* DPC-303: Refactored Organization creation to be an operation

I know we hate operations, but this seems to be work nicely, and doesn't require switching back and forth between FHIR and non-FHIR clients.

* DPC-303: Organization search now takes both a token and an Identifier

Good for us, making things better.

* DPC-303: Basic ProviderRole read/write is working.

We can create practitioner roles and read them back.

* DPC-303: Practitioner role tests are passing.

We can create, search and delete things.

Errors are also handled correctly.

* DPC-303: Removed role update, because that's not really a thing.

* DPC-303: Provider roles are seeding correctly.

* DPC-303: Refactored authentication tests and helpers

Also wired in the Practitioner resource, finally.

* DPC-303: Added Organization principal to creation request

* DPC-415: Added the ability to disable authentication, by injecting static resources that don't block anything.

* DPC-415: We now support Auth annotations in both handlers

* DPC-93: Config overrides are now working.

We need to pass in the config prefix of each application, otherwise the config overrides (which are system properties) are not overridden correctly.

* DPC-303: API Service is now issuing role submission to attribution service

We really have no clean way of injecting extra context into a creation request (except for the crappy header path), so we have to make two API calls, one to create the provider, and a second one to register them with the organization.

* DPC-303: Corrected creation return types and improve search handling

When creating a provider or provider role, we return the correct status codes.

We're also using tags to search for providers associated to a given organization.

* DPC-303: Now returning providers linked to a given organization.

Added a relationship between organizations and providers, by way of the provider_roles table.

* DPC-303: Verify practitioner security

Verified that an organization cannot see another's practitioners.

* DPC-303: Refactored provider test to be more generic

We can now specify which Organizations we want to load as part of an integration test.

* DPC-303: Created secure authentication test

Refactored the base test class to allow us to test secure and non-secure test setups.

* DPC-303: Organization test is passing using full authentication

* DPC-303: Updated test suite to handle refactorings

Test suite is passing with the new changes.

AuthHandlerTest is being flaky, we may have to pull it out.

* DPC-303: Code climate fixes

* DPC-303: Javadoc cleanups.

* DPC-371: PathAuthorizer is more generic

Refactored the authentication/authorization logic to allow more generic resource associating for a given Organization.

Right now, we can look for generic Resource IDs and verify that the Organization has some association to the given resource. We rely on the methods in the Attribution service to do the actual associations.

* DPC-303: Improve security of endpoints.

* DPC-303: Refactored authentication handlers into a factory pattern

Given that we can't pass state between the authenticators and the authorizers, we need to refactor the handlers to create new classes on each auth request.

This is fairly easy to do with Guice but takes a lot of manual switching up of signatures and tests.
The PractitionerResource and AuthenticationTests are passing, but the rest of the suite is mostly disabled.

* DPC-303: Javadoc cleanups and test corrections

Documented and tested.

* DPC-303: Removed NPI constraint from database

Since we can have multiple providers associated with different organizations, we can't have a unique constraint on the NPI column.

* DPC-303: Invalidate system properties cache

We need to reset the config cache when switching between applications.

* DPC-303: Make error prone happy

* DPC-303: One more error prone fix