DPC-293: Refactor attribution (#180)
* DPC-293: Moved current bundle submission to Bundle endpoint

This should clear up space to do a proper Group submission.

* DPC-293: Added initial models and utilities for creating attribution groups

* DPC-293: Roster creation is working

Updated AttributionFHIR test to use newly created Roster group logic.

* DPC-293: The most minimal, stripped down create/fetch operations

Lots of silliness, mostly on my part, but at least it's partially working.

* DPC-293: Group search is now working, and Hibernate is more type-safe

We can now find Rosters by NPI and MBI, which is really helpful.

We're also now using JPA to generate static access names for properties.

* DPC-293: All submission tests are passing.

Patient creation is now idempotent, but the client is really bonkers so we always have to return Created, even when we don't create anything. Because, no idea.

* DPC-293: Roster updating is working correctly.

It's pretty complicated, lots of unnecessary allocations, but I think it's working.

* DPC-293: FHIR attribution test is passing

Refactored base attribution test, it's now working with Groups instead of Bundles.

* DPC-293: Seed Command seems to be working

A ton of refactoring and tweaks to the loading process.

It's awful, it should come out, but it gets us by.

* DPC-293: Expiration test is working.

* DPC-293: Attribution tests are now passing

Everything is green, lots of cleanups remaining, but the code is now passing.

* DPC-293: Added Practitioner $submit operation

We can now submit a bundle of Practitioners, in one fell swoop.

* DPC-293: Patient endpoint now supports $submit operation.

We can now submit Patients in bundles.

* DPC-293: EndToEnd Postman tests are now passing.

We'll need to go back and update everything else, but at least the overall flow works.

* DPC-293: Attribution roster tests are passing

Attribution Service and E2E tests are passing correctly.

* DPC-293: All attribution tests are passing

Removed cascading deletes from the patient entity. We now remove the attributions manually, which is much cleaner and more performant.

We'll keep making these types of changes as we go along.

* DPC-293: Updated mocks for export submission tests

Lots of gross mocking in order to get the HAPI client working correctly.

All of the tests are now passing.

* DPC-293: Group resource now has search/update/delete endpoints

Updated the E2E tests to verify that Group updating works as expected.

A few cleanups on how we handle passing search params back to the attribution service.
It's incredibly cumbersome, but does seem to work.

* DPC-293: Group search now uses Patient ID rather than MBI

This is the correct meaning of the Member query param.

* DPC-293: Error prone and Javadoc cleanups

Some small cleanups to get the build to pass.

* DPC-293: Error prone is now passing

Cleaned up the interaction between JPA and ErrorProne, they're both working now. We have to do two compilation passes, but they're fairly quick.

Also removed the Javadoc warnings for the Jooq code.

A few method and param cleanups to make ErrorProne happy.

* DPC-293: Code Climate cleanups

* DPC-293: Removed unnecessary Roster Engine code

With the new refactor, we no longer need all of the custom code for handling Rosters, so it's been removed.

* DPC-293: Added Operation definitions to FHIR IG

Also updated Capabilities statement to include the new Group endpoint.

* DPC-293: Updated Capabilities statement with new search params

* DPC-293: Attribution Roster validations are passing.

* DPC-293: Final tweaks to FHIR docs

Wired in the final set of tweaks for getting the Attribution profile linked into the FHIR IG and capabilities statement.

Also updated the web docs with the new $submit operations.

* DPC-293: Update missed resource test

* DPC-293: Remove group profiling.

Removing this until DPC-532 is merged.