Updated authorization pattern in Data resource (#190)
We now just require that the user have a valid auth token, we don't do any specific checking of permissions before returning the file. This will probably have to be revisited at a later date.

Also fixed a small bug in the base URL and updated the Demo Command to fully download the outputted files.